added InvoiceRecordPaymentOptions plugin component inside payment reconciliation sheet
